Support floating point power values. - Fix PowerSumming Observer for more than one instance (get rid of static variables)
Keywords: 




git-archimport-id: falk@cs.fau.de--2008-systemoc-intern/SystemC-VPC--multihop--0.6--patch-27

+++ hscd_vpc_Component.cpp
@@ -240,7 +240,7 @@ namespace SystemC_VPC{
       // No '=' character found!
       return false;
     }
-    const int value = atoi(equals + 1);
+    const double value = atof(equals + 1);
     
     if(strncmp(sValue, "IDLE", 4) == 0) {
       powerTable[Component::IDLE] = value;
